requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
mysqlclient==2.2.7
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4

